Separate n8n brief rendering from orchestration
The daily brief service mixed IO orchestration with portfolio summarization and Discord text rendering. This keeps fetch_daily_brief as the stable service entrypoint while moving pure summary/render builders into focused modules that can be tested and evolved independently.\n\nConstraint: /api/n8n/daily-brief response behavior and legacy private helper imports used by tests must remain compatible.\nRejected: Slim app/routers/n8n.py in this branch | router cleanup is a separate lane and would increase merge conflicts.\nConfidence: high\nScope-risk: narrow\nReversibility: clean\nDirective: Keep n8n_daily_brief_service.py responsible for fetching/persistence orchestration; put pure presentation logic in rendering/portfolio helpers.\nTested: uv run --group dev ruff check app/services/n8n_daily_brief_service.py app/services/n8n_daily_brief_portfolio.py app/services/n8n_daily_brief_rendering.py; uv run --group test pytest tests/test_n8n_daily_brief_service.py tests/test_dust_accounting.py tests/test_n8n_daily_brief_formatting.py tests/test_n8n_api.py -q (71 passed)\nNot-tested: Full repository test suite
